requirements.txt
# =============================
# Core Scientific Libraries
# =============================
numpy<2

# =============================
# Image Processing & I/O
# =============================
opencv-python
Pillow
imageio
imagehash
matplotlib
seaborn

# =============================
# Machine Learning & Models
# =============================
torch
torchvision
torchaudio
lpips
open-clip-torch
timm
transformers
pyiqa
spandrel
spandrel_extra_arches
chainner-ext

# =============================
# CLI, UI, and User Input
# =============================
questionary

# =============================
# Audio Feedback
# =============================
pygame
playsound
playsound2

# =============================
# Parallel Processing & Monitoring
# =============================
psutil
joblib
GPUtil

# =============================
# File, Path, and Directory Utilities
# =============================
python-magic==0.4.27
python-magic-bin==0.4.14
libmagic==1.0

# =============================
# Data Compression & Archiving
# =============================
ffmpeg

# =============================
# Jinja2 for Report Templates
# =============================
jinja2

# =============================
# Miscellaneous Utilities
# =============================
pyyaml
pepeline
pepedpid
importlib
gdown

# =============================
# Deduplication & Similarity
# =============================
imagededup

# =============================
# Optional/External Tools
# =============================

# getnative
# ^ first install VapourSynth

# Unsorted/Ungrouped:

pytest